third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-position/position-absolute-in-inline-002-expected.txt

This is a testharness.js-based test.
FAIL absolute inside inline container location should be correct. assert_equals: left matches container expected 69 but got 0
Harness: the test ran to completion.
